requirements.txt
asgiref==3.3.4
Django==3.2.3
django-polymorphic==3.0.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
